requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bootstrap4==0.1.0
breadability==0.1.20
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-heroku==0.3.1
docopt==0.6.2
environ==1.0
gunicorn==20.0.4
idna==2.9
Janome==0.3.10
joblib==0.14.1
lxml==4.5.0
nltk==3.5
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
Pillow==7.1.1
psycopg2==2.8.5
pycountry==19.8.18
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0
sqlparse==0.3.1
sumy==0.8.1
tinysegmenter==0.4
tqdm==4.45.0
urllib3==1.25.8
whitenoise==5.0.1
